cleanup vp
