+++ src/Hyperledger.Aries/Features/OpenID4VC/Vp/Models/AuthorizationRequest.cs
@@ -1,0 +1,9 @@
+using Hyperledger.Aries.Features.Pex.Models;
+
+namespace Hyperledger.Aries.Features.OpenId4Vc.Vp.Models
+{
+    public class AuthorizationRequest
+    {
+        public PresentationDefinition PresentationDefinition { get; set; }
+    }
+}
