+++ src/Hyperledger.Aries/Features/OpenID4VC/Vp/Models/CredentialCandidates.cs
@@ -1,0 +1,12 @@
+using System.Collections.Generic;
+using Hyperledger.Aries.Storage.Models.Interfaces;
+
+namespace Hyperledger.Aries.Features.OpenId4Vc.Vp.Models
+{
+    public class CredentialCandidates
+    {
+        public string InputDescriptorId { get; set; }
+        
+        public List<ICredential> Credentials { get; set; }
+    }
+}
