+++ src/Hyperledger.Aries/Features/Pex/Models/CredentialDescriptor.cs
@@ -1,14 +1,15 @@
-namespace Hyperledger.Aries.Features.Pex.Models;
-
-public class CredentialDescriptor
+namespace Hyperledger.Aries.Features.Pex.Models
 {
-    public string InputDescriptorId { get; set; }
+    public class CredentialDescriptor
+    {
+        public string InputDescriptorId { get; set; }
         
-    public string CredentialId { get; set; }
+        public string CredentialId { get; set; }
         
-    public string Format { get; set; }
+        public string Format { get; set; }
         
-    public string Path { get; set; }
+        public string Path { get; set; }
         
-    public CredentialDescriptor? PathNested { get; set; }
+        public CredentialDescriptor? PathNested { get; set; }
+    }
 }
