+++ src/Hyperledger.Aries/Features/SdJwt/Services/SdJwtCredentialService/ISdJwtCredentialService.cs
@@ -2,8 +2,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Hyperledger.Aries.Agents;
 using Hyperledger.Aries.Features.OpenId4Vc.Vci.Models.Metadata.Issuer;
-using Hyperledger.Aries.Features.OpenID4VP.Models;
-using Hyperledger.Aries.Features.OpenID4VP.Services;
+using Hyperledger.Aries.Features.OpenId4Vc.Vp.Models;
 using Hyperledger.Aries.Features.Pex.Models;
 using Hyperledger.Aries.Features.SdJwt.Models;
 using Hyperledger.Aries.Features.SdJwt.Models.Records;
