+++ src/Hyperledger.Aries/Hyperledger.Aries.csproj
@@ -20,7 +20,4 @@
         <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.25.1" />
         <PackageReference Include="System.Reactive.Linq" Version="$(SystemReactiveLinqVersion)" />
     </ItemGroup>
-    <ItemGroup>
-      <Folder Include="Model\" />
-    </ItemGroup>
 </Project>
