+++ src/Hyperledger.Aries/Storage/Models/Interfaces/ICredential.cs
@@ -1,0 +1,6 @@
+namespace Hyperledger.Aries.Storage.Models.Interfaces
+{
+    public interface ICredential
+    {
+    }
+}
