--- src/Hyperledger.Aries/Storage/Records/CredentialRecordBase.cs
@@ -1,5 +1,0 @@
-namespace Hyperledger.Aries.Storage;
-
-public interface ICredential
-{
-}
